Rename Utilities to Configuration [Persistence]
